renew

+++ app/src/main/java/com/amdc/firebasetest/MessageAdapter.java
@@ -34,7 +34,8 @@ import static android.os.Environment.DIRECTORY_DOWNLOADS;
 
 public class MessageAdapter extends RecyclerView.Adapter<MessageAdapter.MessageViewHolder> {
     private List<Messages> userMessagesList;
-    private FirebaseAuth mAuth;
+    String receiverImage, userToName, userToImage, toUserID;
+//    private FirebaseAuth mAuth;
     private StorageReference storageReference, reference;
 
     MessageAdapter(List<Messages> userMessagesList) {
@@ -63,23 +64,35 @@ public class MessageAdapter extends RecyclerView.Adapter<MessageAdapter.MessageV
     @Override
     public MessageViewHolder onCreateViewHolder(@NonNull ViewGroup viewGroup, int i) {
         View view = LayoutInflater.from(viewGroup.getContext()).inflate(R.layout.custom_messages_layout, viewGroup, false);
-        mAuth = FirebaseAuth.getInstance();
         return new MessageViewHolder(view);
     }
 
     @SuppressLint("SetTextI18n")
     @Override
     public void onBindViewHolder(@NonNull final MessageViewHolder messageViewHolder, final int position) {
-        String messageSenderId = Objects.requireNonNull(mAuth.getCurrentUser()).getUid();
+        Context context = messageViewHolder.itemView.getContext();
+        String messageSenderId = Objects.requireNonNull(FirebaseAuth.getInstance().getCurrentUser()).getUid();
         Messages messages = userMessagesList.get(position);
         String fromUserID = messages.getFrom();
+        toUserID = messages.getTo();
         String fromMessageType = messages.getType();
-        DatabaseReference usersRef = FirebaseDatabase.getInstance().getReference().child("Users").child(fromUserID);
-        usersRef.addValueEventListener(new ValueEventListener() {
+        DatabaseReference usersFromRef = FirebaseDatabase.getInstance().getReference().child("Users").child(fromUserID);
+        DatabaseReference usersToRef = FirebaseDatabase.getInstance().getReference().child("Users").child(toUserID);
+        usersToRef.addValueEventListener(new ValueEventListener() {
+            @Override
+            public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
+                userToName = (String) dataSnapshot.child("name").getValue();
+                userToImage = (String) dataSnapshot.child("image").getValue();
+            }
+            @Override
+            public void onCancelled(@NonNull DatabaseError databaseError) { }
+        });
+
+        usersFromRef.addValueEventListener(new ValueEventListener() {
             @Override
             public void onDataChange(@NonNull DataSnapshot dataSnapshot) {
                 if (dataSnapshot.hasChild("image")) {
-                    String receiverImage = Objects.requireNonNull(dataSnapshot.child("image").getValue()).toString();
+                    receiverImage = Objects.requireNonNull(dataSnapshot.child("image").getValue()).toString();
                     Picasso.get().load(receiverImage).placeholder(R.drawable.profile_image).into(messageViewHolder.receiverProfileImage);
                 }
             }
@@ -182,36 +195,36 @@ public class MessageAdapter extends RecyclerView.Adapter<MessageAdapter.MessageV
                     case "doc":
                     case "docx": {
                         CharSequence[] options = new CharSequence[] {"Download and View this Document", "Download this Document", "Delete From Everywhere", "Delete Only From Me", "Cancel"};
-                        AlertDialog.Builder builder = new AlertDialog.Builder(messageViewHolder.itemView.getContext());
+                        AlertDialog.Builder builder = new AlertDialog.Builder(context);
                         builder.setTitle("Download/Delete Message").setIcon(R.drawable.file);
                         builder.setItems(options, (dialogInterface, i) -> {
                             if (i == 0) {
                                 String fileNameForDownload = userMessagesList.get(position).getMessageID() + "." + userMessagesList.get(position).getType(); //get name file from storage
                                 String fileName = Uri.parse(userMessagesList.get(position).getName()).toString(); //get name file from user into firebase (getLastPathSegment())
-                                Intent intent = new Intent(messageViewHolder.itemView.getContext(), FilesViewerActivity.class);
+                                Intent intent = new Intent(context, FilesViewerActivity.class);
                                 intent.putExtra("fileNameForDownload", fileNameForDownload);
                                 intent.putExtra("fileName", fileName);
-                                messageViewHolder.itemView.getContext().startActivity(intent);
+                                context.startActivity(intent);
                             } else if (i == 1) { // download file
                                 String fileNameForDownload = userMessagesList.get(position).getMessageID() + "." + userMessagesList.get(position).getType(); //get name file from storage
                                 storageReference = FirebaseStorage.getInstance().getReference();
                                 reference = storageReference.child("Document Files").child(fileNameForDownload);
                                 reference.getDownloadUrl().addOnSuccessListener(uri -> {
                                     String fileName = Uri.parse(userMessagesList.get(position).getName()).toString(); //get name file from user into firebase (getLastPathSegment())
-                                    DownloadManager downloadManager = (DownloadManager) messageViewHolder.itemView.getContext().getSystemService(Context.DOWNLOAD_SERVICE);
+                                    DownloadManager downloadManager = (DownloadManager) context.getSystemService(Context.DOWNLOAD_SERVICE);
                                     DownloadManager.Request request = new DownloadManager.Request(uri);
                                     request.setNotificationVisibility(DownloadManager.Request.VISIBILITY_VISIBLE_NOTIFY_COMPLETED);
-                                    request.setDestinationInExternalFilesDir(messageViewHolder.itemView.getContext(), DIRECTORY_DOWNLOADS, fileName); //reference.getName()
+                                    request.setDestinationInExternalFilesDir(context, DIRECTORY_DOWNLOADS, fileName); //reference.getName()
                                     Objects.requireNonNull(downloadManager).enqueue(request);
-                                }).addOnFailureListener(e -> Toast.makeText(messageViewHolder.itemView.getContext(), "Error download", Toast.LENGTH_SHORT).show());
+                                }).addOnFailureListener(e -> Toast.makeText(context, "Error download", Toast.LENGTH_SHORT).show());
                             } if (i == 2) {
                                 deleteMessageForEveryOne(position, messageViewHolder);
-                                Intent intent = new Intent(messageViewHolder.itemView.getContext(), MainActivity.class);
-                                messageViewHolder.itemView.getContext().startActivity(intent);
+                                Intent intent = new Intent(context, MainActivity.class);
+                                context.startActivity(intent);
                             } if (i == 3) {
                                 deleteSentMessage(position, messageViewHolder);
-                                Intent intent = new Intent(messageViewHolder.itemView.getContext(), MainActivity.class);
-                                messageViewHolder.itemView.getContext().startActivity(intent);
+                                Intent intent = new Intent(context, MainActivity.class);
+                                context.startActivity(intent);
                             }
                         });
                         builder.show();
@@ -219,20 +232,23 @@ public class MessageAdapter extends RecyclerView.Adapter<MessageAdapter.MessageV
                     }
                     case "text": { // sender messages for text
                         CharSequence[] options = new CharSequence[]{"Copy text message", "Delete From Everywhere", "Delete Only From Me", "Cancel"};
-                        AlertDialog.Builder builder = new AlertDialog.Builder(messageViewHolder.itemView.getContext());
+                        AlertDialog.Builder builder = new AlertDialog.Builder(context);
                         builder.setTitle("Cory/Delete Message").setIcon(R.drawable.delete);
                         builder.setItems(options, (dialogInterface, i) -> {
                             if (i == 0) {
-                                ((ClipboardManager) Objects.requireNonNull(messageViewHolder.itemView.getContext().getSystemService(Context.CLIPBOARD_SERVICE))).setText(messages.getMessage());
-                                Toast.makeText(messageViewHolder.itemView.getContext(),"Copied to clipboard",Toast.LENGTH_SHORT).show();
+                                ((ClipboardManager) Objects.requireNonNull(context.getSystemService(Context.CLIPBOARD_SERVICE))).setText(messages.getMessage());
+                                Toast.makeText(context,"Copied to clipboard",Toast.LENGTH_SHORT).show();
                             } if (i == 1) {
                                 deleteMessageForEveryOne(position, messageViewHolder);
-                                Intent intent = new Intent(messageViewHolder.itemView.getContext(), MainActivity.class);
-                                messageViewHolder.itemView.getContext().startActivity(intent);
+//                                Intent chatIntent = new Intent(context, ChatActivity.class);
+//                                chatIntent.putExtra("visit_user_id", toUserID);
+//                                chatIntent.putExtra("visit_user_name", userToName);
+//                                chatIntent.putExtra("visit_image", userToImage);
+//                                context.startActivity(chatIntent);
                             } if (i == 2) {
                                 deleteSentMessage(position, messageViewHolder);
-                                Intent intent = new Intent(messageViewHolder.itemView.getContext(), MainActivity.class);
-                                messageViewHolder.itemView.getContext().startActivity(intent);
+                                Intent intent = new Intent(context, MainActivity.class);
+                                context.startActivity(intent);
                             }
                         });
                         builder.show();
@@ -240,21 +256,21 @@ public class MessageAdapter extends RecyclerView.Adapter<MessageAdapter.MessageV
                     }
                     case "image": { // sender messages for image
                         CharSequence[] options = new CharSequence[]{"View This Image", "Delete From Everywhere", "Delete Only From Me", "Cancel"};
-                        AlertDialog.Builder builder = new AlertDialog.Builder(messageViewHolder.itemView.getContext());
+                        AlertDialog.Builder builder = new AlertDialog.Builder(context);
                         builder.setTitle("View/Delete Message").setIcon(R.drawable.file);
                         builder.setItems(options, (dialogInterface, i) -> {
                             if (i == 0) {
-                                Intent intent = new Intent(messageViewHolder.itemView.getContext(), ImageViewerActivity.class);
+                                Intent intent = new Intent(context, ImageViewerActivity.class);
                                 intent.putExtra("url", userMessagesList.get(position).getMessage());
-                                messageViewHolder.itemView.getContext().startActivity(intent);
+                                context.startActivity(intent);
                             } if (i == 1) {
                                 deleteMessageForEveryOne(position, messageViewHolder);
-                                Intent intent = new Intent(messageViewHolder.itemView.getContext(), MainActivity.class);
-                                messageViewHolder.itemView.getContext().startActivity(intent);
+                                Intent intent = new Intent(context, MainActivity.class);
+                                context.startActivity(intent);
                             } if (i == 2) {
                                 deleteSentMessage(position, messageViewHolder);
-                                Intent intent = new Intent(messageViewHolder.itemView.getContext(), MainActivity.class);
-                                messageViewHolder.itemView.getContext().startActivity(intent);
+                                Intent intent = new Intent(context, MainActivity.class);
+                                context.startActivity(intent);
                             }
                         });
                         builder.show();
@@ -397,16 +413,27 @@ public class MessageAdapter extends RecyclerView.Adapter<MessageAdapter.MessageV
                         .removeValue().addOnCompleteListener(task1 -> { //delete sms with name file from receive messages
                     if(task1.isSuccessful()) {
                         storageReference = FirebaseStorage.getInstance().getReference();
-                        if (userMessagesList.get(position).getType().equals("image")) { //for image
-                            String fileNameForDownload = userMessagesList.get(position).getMessageID() + ".jpg"; //name file form storage
-                            storageReference.child("Image Files").child(fileNameForDownload).delete().addOnCompleteListener(task2 -> {
-                                if (task2.isSuccessful()) Toast.makeText(holder.itemView.getContext(), "Deleted Successfully.", Toast.LENGTH_SHORT).show();
-                            });
-                        } else { //for files
-                            String fileNameForDownload = userMessagesList.get(position).getMessageID() + "." + userMessagesList.get(position).getType(); //name file form storage
-                            storageReference.child("Document Files").child(fileNameForDownload).delete().addOnCompleteListener(task2 -> { //delete file from storage
-                                if (task2.isSuccessful()) Toast.makeText(holder.itemView.getContext(), "Deleted Successfully.", Toast.LENGTH_SHORT).show();
-                            });
+                        if (userMessagesList.get(position).getType().equals("text")) {
+                            Toast.makeText(holder.itemView.getContext(), "Message Deleted Successfully.", Toast.LENGTH_SHORT).show();
+                            Intent chatIntent = new Intent(holder.itemView.getContext(), ChatActivity.class);
+                            chatIntent.putExtra("visit_user_id", toUserID);
+                            chatIntent.putExtra("visit_user_name", userToName);
+                            chatIntent.putExtra("visit_image", userToImage);
+                            holder.itemView.getContext().startActivity(chatIntent);
+                        } else {
+                            if (userMessagesList.get(position).getType().equals("image")) { //for image
+                                String fileNameForDownload = userMessagesList.get(position).getMessageID() + ".jpg"; //name file form storage
+                                storageReference.child("Image Files").child(fileNameForDownload).delete().addOnCompleteListener(task2 -> {
+                                    if (task2.isSuccessful())
+                                        Toast.makeText(holder.itemView.getContext(), "Deleted Successfully.", Toast.LENGTH_SHORT).show();
+                                });
+                            } else { //for files
+                                String fileNameForDownload = userMessagesList.get(position).getMessageID() + "." + userMessagesList.get(position).getType(); //name file form storage
+                                storageReference.child("Document Files").child(fileNameForDownload).delete().addOnCompleteListener(task2 -> { //delete file from storage
+                                    if (task2.isSuccessful())
+                                        Toast.makeText(holder.itemView.getContext(), "Deleted Successfully.", Toast.LENGTH_SHORT).show();
+                                });
+                            }
                         }
                     }
                 });
